refactor: ♻️ project detail sidebar contributions card
