+++ components/partials/project/[id]/ProjectSidebar.tsx
@@ -1,34 +1,26 @@
 import { Button, Card, Icon, Stack, Text } from "@bbtgnn/polaris-interfacer";
-import { ListBoxes, MagicWand, ParentChild } from "@carbon/icons-react";
+import { ParentChild } from "@carbon/icons-react";
 import { LinkMinor, PlusMinor } from "@shopify/polaris-icons";
 import AddStar from "components/AddStar";
 import BrDisplayUser from "components/brickroom/BrDisplayUser";
 import { useProject } from "components/layout/FetchProjectLayout";
 import OshTool from "components/Osh";
-import ProjectContributors from "components/ProjectContributors";
 import { ProjectType } from "components/types";
 import WatchButton from "components/WatchButton";
 import { useAuth } from "hooks/useAuth";
 import useStorage from "hooks/useStorage";
 import { ResourceProposalsQuery } from "lib/types";
 import { useTranslation } from "next-i18next";
-import { useRouter } from "next/router";
-import { Dispatch, SetStateAction, useState } from "react";
+import { useState } from "react";
 import ActionsCard from "./sidebar/ActionsCard";
+import ContributionsCard from "./sidebar/ContributionsCard";
 
-interface Props {
-  contributions: ResourceProposalsQuery;
-  setSelected: Dispatch<SetStateAction<number>>;
-}
-
-export default function ProjectSidebar(props: Props) {
+export default function ProjectSidebar() {
   const { t } = useTranslation("common");
-  const { contributions, setSelected } = props;
-  const { project } = useProject();
+  const { project, setSelected } = useProject();
   const { user } = useAuth();
   const { getItem, setItem } = useStorage();
   const [inList, setInList] = useState<boolean>(false);
-  const router = useRouter();
 
   const handleCollect = () => {
     const _list = getItem("projectsCollected");
@@ -43,10 +35,6 @@ export default function ProjectSidebar(props: Props) {
     }
   };
   const isDesign = project.conformsTo?.name === ProjectType.DESIGN;
-  const licenses = project.metadata?.licenses?.length > 0 && project?.metadata?.licenses;
-  const design = project.metadata?.design;
-  const declarations = project.metadata?.declarations;
-  const haveDeclarations = declarations?.recyclable === "yes" || declarations?.repairable === "yes";
 
   return (
     <div className="lg:col-span-1 order-first lg:order-last">
@@ -79,41 +67,7 @@ export default function ProjectSidebar(props: Props) {
       </Card>
 
       <ActionsCard />
-      {/* Contributions */}
-      <Card sectioned>
-        <Stack vertical>
-          <Text as="h2" variant="headingMd">
-            {t("Contributions")}
-          </Text>
-          <Text color="success" as="p" variant="bodyMd" id="sidebar-contributors">
-            {t("{{contributors}} contributors", { contributors: project.metadata.contributors?.length || 0 })}
-          </Text>
-          <ProjectContributors projectNode={project} />
-          <Text color="success" as="p" variant="bodyMd">
-            {t("{{contributions}} contributions", { contributions: contributions?.proposals.edges.length })}
-          </Text>
-          <Button
-            id="contribute"
-            icon={<MagicWand />}
-            size="large"
-            fullWidth
-            primary
-            onClick={() => router.push(`/create/contribution/${project.id}`)}
-          >
-            {t("Make a contribution")}
-          </Button>
-          <Button
-            id="seeContributions"
-            icon={<ListBoxes />}
-            size="large"
-            fullWidth
-            monochrome
-            onClick={() => setSelected(4)}
-          >
-            {t("All contributions")}
-          </Button>
-        </Stack>
-      </Card>
+      <ContributionsCard />
       {/* Relations */}
       <Card sectioned>
         <Stack vertical spacing="loose">
